requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.0.1
requests==2.31.0
aiofiles==23.2.1
pydantic==2.5.0
openai==1.3.5
google-cloud-vision==3.4.0
boto3==1.28.0
python-dateutil==2.8.2